Temporarily rename directory
