+++ cms/dynamic_content/global_filter/card.py
@@ -1,15 +1,9 @@
-import datetime
-
-from django.core.exceptions import ValidationError
-from django.forms.utils import ErrorList
 from wagtail import blocks, fields
 
 from cms.dynamic_content import help_texts
-from cms.dynamic_content.global_filter.filters import (
-    GEOGRAPHY_TYPE_FIELDS,
-    DataFilter,
-    ThresholdsFilter,
-)
+from cms.dynamic_content.global_filter.filter_types.thresholds import ThresholdsFilter
+from cms.dynamic_content.global_filter.filter_types.time_range import TimeRangeBlock
+from cms.dynamic_content.global_filter.filters import GEOGRAPHY_TYPE_FIELDS, DataFilter
 
 MINIMUM_FILTER_ROWS_COUNT = 1
