+++ cms/dynamic_content/global_filter/filter_types/time_range.py
