+++ tests/unit/cms/dynamic_content/global_filter/filters/test_thresholds.py
@@ -1,11 +1,10 @@
-import datetime
 from unittest import mock
 
 import pytest
 from decimal import Decimal
 from wagtail.blocks import StructBlockValidationError, StructBlock
 
-from cms.dynamic_content.global_filter.filters.thresholds import (
+from cms.dynamic_content.global_filter.filter_types.thresholds import (
     ThresholdFilterElement,
     ThresholdsFilter,
 )
